+++ data-raw/achilles.R
@@ -1,0 +1,5 @@
+library(usethis)
+
+ach <- Achilles(rowRanges = FALSE, colData = FALSE)
+ach <- ach[seq_len(100L), seq_len(50L)]
+use_data(ach, overwrite = TRUE)